Move views to separate view fodler
For better file structure
